+++ crates/zng-unique-id/src/lib.rs
@@ -83,7 +83,7 @@ macro_rules! unique_id_32 {
                 /// a lazy static that instantiates the ID.
                 ///
                 /// # Exhaustion Handling
-                /// 
+                ///
                 /// If more IDs are generated them `u32::MAX` an error is logged, the internal counter is reset and ids are reused.
                 $vis struct $Type $(< $T $(:($($bounds)+))? >)? $(: $ParentId)? ;
             }
@@ -138,7 +138,7 @@ macro_rules! unique_id_64 {
                 /// a lazy static that instantiates the ID.
                 ///
                 /// # Exhaustion Handling
-                /// 
+                ///
                 /// If more IDs are generated them `u64::MAX` an error is logged, the internal counter is reset and ids are reused.
                 $vis struct $Type $(< $T $(:($($bounds)+))? >)? $(: $ParentId)? ;
             }